requirements.txt
# Core dependencies
rich>=13.0.0
textual>=0.41.0
click>=8.0.0
pydantic>=2.0.0

# Terraform parsing and analysis
python-hcl2>=4.0.0
jinja2>=3.0.0

# Natural language processing
openai>=1.0.0
tiktoken>=0.5.0

# LangChain for conversation management and RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0

# File system and process management
watchdog>=3.0.0
psutil>=5.9.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# Utilities
colorama>=0.4.6
python-dotenv>=1.0.0
